attempt1/requirements.txt
newsapi-python>=0.2.6
tweepy>=4.14.0
PyYAML>=6.0
pymongo>=4.5.0
nltk>=3.8
spacy>=3.6.0